requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
Pillow==9.0.1
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
requests==2.27.1
urllib3==1.26.8
